+++ tests/testthat/test-carehomes-vaccination.R
@@ -1693,6 +1693,15 @@ test_that("check_rel_param allows sensible inputs", {
 })
 
 
+test_that("check_rel_param allows <1 probs for second infection", {
+  expect_silent(
+    check_rel_param(array(c(1, 1, 0.5, 0.5), c(1, 4, 1)), "rel_param"))
+  expect_error(
+    check_rel_param(array(c(1, 0.5, 0.5, 0.5), c(1, 4, 1)), "rel_param"),
+                    "must be 1 for first infection")
+})
+
+
 test_that("build_rel_param rejects wrong dimension or out of bound inputs", {
   expect_error(
     build_rel_param(