+++ NEWS.md
@@ -1,3 +1,7 @@
+# sircovid 0.10.50
+
+* rel vacc params can now be less than one in unvaccinated classes for second infections (strains 3 and 4)
+
 # sircovid 0.10.49
 
 * Add `lag_groups` and `lag_days` parameters to `vaccine_schedule_future` to manually add lags for given age groups to the vaccine schedule